fix: separate 'authCheck' from 'auth' file
It bloats the bundle size because of the 'verify' function from
'jsonwebtoken' being imported into the client, when normally it only
runs in the Node server (via getServerSideProps)
