+++ client/src/lib/auth.js
@@ -1,24 +1,7 @@
 import Router from 'next/router';
 import ky from 'ky-universal';
-import cookie from 'cookie';
-import { verify } from 'jsonwebtoken';
 import HOST from '@/lib/host';
 const api = `${HOST}/api/auth`;
-const secret = process.env.SECRET;
-
-const authCheck = async (ctx) => {
-  const noAuth = { username: null, role: null };
-
-  const reqCookie = ctx.req.headers.cookie;
-  if (!reqCookie) return noAuth; // No token
-
-  const cookies = cookie.parse(reqCookie);
-  const payload = verify(cookies.auth, secret);
-  if (!payload) return noAuth; // Invalid token
-
-  const { username, role } = payload;
-  return (username && role) ? { username, role } : noAuth;
-};
 
 export const signup = async (data) => {
   try {
@@ -63,5 +46,3 @@ export const logout = async () => {
   await ky.post(`${api}/logout`, { json: { 'message': 'Log out' }});
   Router.reload();
 };
-
-export default authCheck;
